+++ lib/haskell/natural4/src/LS/XPile/Maude/Maude.hs
@@ -1,0 +1,6 @@
+module LS.XPile.Maude.Maude
+  ( rules2maudeStr,
+  )
+where
+
+import LS.XPile.Maude.Rules (rules2maudeStr)
